dom/jsurl/test/test_bug1382035-1.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1382035
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1382035</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 1382035 **/
  addLoadEvent(function() {
    is(frames[0].document.documentElement.textContent, "",
       "Should not navigate to a stringified Promise");
    SimpleTest.finish();
  });

  SimpleTest.waitForExplicitFinish();
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1382035">Mozilla Bug 1382035</a>
<p id="display"></p>
<div id="content" style="display: none">
  <iframe src="javascript: Promise.resolve()">
  </iframe>
</div>
<pre id="test">
</pre>
</body>
</html>
